testing-framework/requirements.txt
# AI Testing Framework Dependencies

# Claude API
anthropic>=0.40.0

# HTTP clients
httpx>=0.27.0

# Data formats
pyyaml>=6.0
pydantic>=2.0.0

# Console UI
rich>=13.7.0

# Templates
jinja2>=3.1.2

# Date/time utilities
python-dateutil>=2.8.2
